DDOS

+++ src/app/api/data/removeData/[id]/route.ts
@@ -1,10 +1,16 @@
 import { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 import { PrismaClient } from '@/generated/prisma';
+import { getToken } from 'next-auth/jwt';
 
 export async function DELETE(req: NextRequest, { params }: { params: any }) {
   const prisma = new PrismaClient();
   const { id } = params;
+  const token = await getToken({ req });
+
+  if (!token) {
+    return null;
+  }
 
   try {
     await prisma.user.delete({ where: { id } });
